Move files and fix package
