+++ packages/compass-e2e-tests/.prettierignore
@@ -4,8 +4,3 @@ fixtures
 .nyc_output
 coverage
 hadron-build-info.json
-
-# Ignoring sandboxes (created per test run)
-.smoke-sandboxes/
-# Cache of downloaded binaries
-.smoke-downloads/
