+++ packages/compass-smoke-tests/.eslintrc.js
@@ -3,6 +3,14 @@ module.exports = {
   extends: ['@mongodb-js/eslint-config-compass'],
   parserOptions: {
     tsconfigRootDir: __dirname,
-    project: ['./tsconfig-lint.json'],
+    project: ['./tsconfig.json'],
   },
+  overrides: [
+    {
+      files: ['**/*.ts'],
+      rules: {
+        'no-console': 0,
+      },
+    },
+  ],
 };
