+++ packages/compass-smoke-tests/.gitignore
@@ -1,7 +1,3 @@
-.nyc_output
-dist
-coverage
-
 # Sandboxes (created per test run)
 .sandboxes/
 # Cache of downloaded files
