--- packages/compass-smoke-tests/.mocharc.js
@@ -1,1 +1,0 @@
-module.exports = require('@mongodb-js/mocha-config-compass');
