+++ packages/compass-smoke-tests/package.json
@@ -16,13 +16,10 @@
     "type": "git",
     "url": "https://github.com/mongodb-js/compass.git"
   },
-  "files": [
-    "dist"
-  ],
+  "compass:exports": {},
   "license": "SSPL",
   "scripts": {
     "start": "ts-node src/cli.ts",
-    "bootstrap": "npm run compile",
     "typecheck": "tsc -p tsconfig.json",
     "eslint": "eslint",
     "prettier": "prettier",
@@ -33,12 +30,16 @@
     "reformat": "npm run eslint . -- --fix && npm run prettier -- --write ."
   },
   "devDependencies": {
+    "@types/node": "^20",
     "@mongodb-js/eslint-config-compass": "^1.2.1",
     "@mongodb-js/prettier-config-compass": "^1.1.1",
     "@mongodb-js/tsconfig-compass": "^1.1.1",
     "depcheck": "^1.4.1",
     "eslint": "^7.25.0",
+    "hadron-build": "^25.6.1",
+    "lodash": "^4.17.21",
     "prettier": "^2.7.1",
-    "typescript": "^5.0.4"
+    "typescript": "^5.0.4",
+    "yargs": "^17.7.2"
   }
 }
