+++ packages/compass-smoke-tests/src/installers/mac-dmg.ts
@@ -2,17 +2,27 @@ import path from 'node:path';
 import fs from 'node:fs';
 
 import type { InstalledAppInfo, InstallablePackage } from './types';
-import { execute } from './helpers';
+import { execute } from '../execute';
 
-export function installMacZIP({
+export function installMacDMG({
   appName,
   filepath,
   destinationPath,
 }: InstallablePackage): InstalledAppInfo {
   const appFilename = `${appName}.app`;
   const appPath = path.resolve(destinationPath, appFilename);
+  const volumePath = `/Volumes/${appName}`;
 
-  execute('ditto', ['-xk', filepath, destinationPath]);
+  execute('hdiutil', ['attach', filepath]);
+
+  try {
+    fs.cpSync(path.resolve(volumePath, appFilename), appPath, {
+      recursive: true,
+      verbatimSymlinks: true,
+    });
+  } finally {
+    execute('hdiutil', ['detach', volumePath]);
+  }
 
   // TODO: Consider instrumenting the app to use a settings directory in the sandbox
   // TODO: Move this somewhere shared between mac installers
