+++ packages/compass-smoke-tests/src/installers/types.ts
