+++ packages/compass-smoke-tests/src/packages.ts
