+++ packages/compass-smoke-tests/tsconfig.json
@@ -2,7 +2,8 @@
   "extends": "@mongodb-js/tsconfig-compass/tsconfig.common.json",
   "compilerOptions": {
     "noEmit": true,
-    "types": ["node"]
+    "types": ["node"],
+    "lib": ["ES2021"]
   },
   "include": ["src/**/*"]
 }
